+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.2.4 - 2024-09-24
+
+### Added
+
+- [Add Federated Prometheus Monitoring Provider](https://sedaiengineering.github.io/sedai-sdk-python/sedai/monitoring_provider.html#add_federated_prometheus_monitoring)
+- [Agentless Kubernetes setup examples](./examples/self_managed_setup/setup_cluster.py)
+
 ## 1.2.3 - 2024-08-20
 
 ### Added